Core/Formulas: implemented 90% experience gain reduction for levels lower than max in mobs expansion. Also fixed content level function for cataclysm zones on old maps

+++ src/server/game/Groups/Group.cpp
@@ -682,7 +682,7 @@ void Group::GroupLoot(Loot *loot, WorldObject* pLootedObject)
             {
                 r->setLoot(loot);
                 r->itemSlot = itemSlot;
-                if (item->DisenchantID && m_maxEnchantingLevel >= item->RequiredDisenchantSkill)
+                if (item->DisenchantID && int32(m_maxEnchantingLevel) >= item->RequiredDisenchantSkill)
                     r->rollVoteMask |= ROLL_FLAG_TYPE_DISENCHANT;
 
                 loot->items[itemSlot].is_blocked = true;
@@ -770,7 +770,7 @@ void Group::NeedBeforeGreed(Loot *loot, WorldObject* pLootedObject)
             {
                 r->setLoot(loot);
                 r->itemSlot = itemSlot;
-                if (item->DisenchantID && m_maxEnchantingLevel >= item->RequiredDisenchantSkill)
+                if (item->DisenchantID && int32(m_maxEnchantingLevel) >= item->RequiredDisenchantSkill)
                     r->rollVoteMask |= ROLL_FLAG_TYPE_DISENCHANT;
 
                 if (item->Flags2 & ITEM_FLAGS_EXTRA_NEED_ROLL_DISABLED)
